requirements.txt
absl-py==1.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==21.4.0
backcall==0.2.0
bleach==4.1.0
cached-property==1.5.2
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.4
cycler==0.11.0
Cython==0.29.28
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
delaunay==1.0.3
docopt==0.6.2
entrypoints==0.4
fonttools==4.28.5
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.2
importlib-resources==5.4.0
intel-openmp==2022.0.3
ipykernel==6.9.1
ipython==7.32.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
kiwisolver==1.3.2
Markdown==3.3.7
MarkupSafe==2.1.0
matlabengineforpython===R2018a
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.11
nbconvert==6.4.2
nbconvert_reportlab==0.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.21.5
oauthlib==3.2.0
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pdfrw==0.4
pickleshare==0.7.5
Pillow==9.0.0
pipreqs==0.4.11
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygame==2.1.2
Pygments==2.11.2
PyOpenGL==3.1.5
pyparsing==3.0.7
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
PyQt5-stubs==5.15.2.0
pyqt5-tools==5.15.4.3.2
PyQt5Designer==5.14.1
pyrsistent==0.18.1
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pywin32==303
pywinpty==2.0.2
PyYAML==6.0
pyzmq==22.3.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
qtconsole==5.2.2
QtPy==2.0.1
reportlab==3.6.8
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.13.1
testpath==0.6.0
threadpoolctl==3.1.0
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchvision==0.11.3+cu113
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.5.2
yarg==0.1.9
zipp==3.7.0
